changed int to float on salary and total cost

+++ house_hunting.py
@@ -3,9 +3,9 @@ current_savings = 0
 r = .04
 portion_down_payment = .25
 
-annual_salary = int(input('Enter your annual salary: '))
+annual_salary = float(input('Enter your annual salary: '))
 portion_saved = float(input('Enter the percent of your salary to save, as a decimal: '))
-total_cost = int(input('Enter the cost of your dream home: '))
+total_cost = float(input('Enter the cost of your dream home: '))
 
 count = 0
 while current_savings < total_cost*portion_down_payment:
